# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "UiTdatabank v3",
        "description": "Version 3 of the UiTdatabank, a central database of cultural offers in the Flanders region.\n\nMost operations require that you authenticate first with your UiTID.",
        "version": "3.0.0",
        "contact": {
            "name": "CultuurNet Vlaanderen vzw",
            "url": "http://www.cultuurnet.be",
            "email": "info@uitdatabank.be"
        }
    },
    "basePath": "/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json",
        "application/ld+json",
        "application/xml"
    ],
    "paths": {
        "/api/1.0/event": {
            "post": {
                "summary": "Create a new cultural event.",
                "operationId": "createEvent",
                "produces": [
                    "application/ld+json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing the necessary information to create a new cultural event.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/eventCreationData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the created event.",
                        "schema": {
                            "properties": {
                                "eventId": {
                                    "type": "string",
                                    "format": "uuid",
                                    "description": "The unique identifier of the cultural event.",
                                    "example": "0823f57e-a6bd-450a-b4f5-8459b4b11043"
                                },
                                "url": {
                                    "type": "string",
                                    "format": "url",
                                    "description": "The URL of the cultural event.",
                                    "example": "http://culudb-silex.dev:8080/event/0823f57e-a6bd-450a-b4f5-8459b4b11043"
                                }
                            },
                            "required": [
                                "eventId",
                                "url"
                            ]
                        }
                    }
                },
                "tags": [
                    "Event"
                ]
            }
        },
        "/event/{cdbid}": {
            "get": {
                "summary": "View the details of a specific cultural event.",
                "operationId": "getEvent",
                "produces": [
                    "application/ld+json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the cultural event.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An event detail.",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    }
                },
                "tags": [
                    "Event"
                ]
            }
        },
        "/event/{cdbid}/history": {
            "get": {
                "summary": "View the change log / history of a specific cultural event.",
                "operationId": "getEventHistory",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the cultural event.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The event history.",
                        "schema": {
                            "$ref": "#/definitions/EventHistory"
                        }
                    }
                },
                "tags": [
                    "Event"
                ]
            }
        },
        "/event/{cdbid}/permission": {
            "get": {
                "summary": "Whether the current user has permission to update.",
                "operationId": "hasPermission",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the event.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "has permission.",
                        "schema": {
                            "properties": {
                                "hasPermission": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "required": [
                                "hasPermission"
                            ]
                        }
                    }
                },
                "tags": [
                    "Event"
                ]
            }
        },
...
        "/event/{cdbid}/{lang}/description": {
            "post": {
                "summary": "Translate the description of a cultural event.",
                "description": "Authentication is required.",
                "operationId": "translateEventDescription",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the cultural event.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "The language of the translation.",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "fr",
                            "de",
                            "en"
                        ]
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing the description in the given language for the given cultural event.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postObjectWithEventDescription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Event"
                ]
            }
        },
...
        "/event/{cdbid}/images": {
            "post": {
                "summary": "Add an MediaObject to the Event",
                "description": "Authentication is required",
                "operationId": "EventAddImage",
                "produces": [
                    "application/ld+json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The universally unique identifier of the event object.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "The UUID of the image to add",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/addImageToOffer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    },
                    "400": {
                        "description": "Media with id: {id} not found."
                    }
                },
                "tags": [
                    "Event"
                ]
            }
        },
...
        "/event/{eventId}/images/{mediaObjectId}": {
            "post": {
                "summary": "Update the info of an event image.",
                "description": "Authentication is required.",
                "operationId": "updateEventImage",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "The unique identifier of the event.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "mediaObjectId",
                        "in": "path",
                        "description": "The unique identifier of the image media object.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "An object containing the image information that has to be changed.",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/imageUpdateInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Info on the command that was queued.",
                        "schema": {
                            "$ref": "#/definitions/command"
                        }
                    }
                },
                "tags": [
                    "Event",
                    "Media"
                ]
            }
        },
...
        "/place/{cdbid}/permission": {
            "get": {
                "summary": "Whether the current user has permission to update.",
                "operationId": "hasPermission",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cdbid",
                        "in": "path",
                        "description": "The unique identifier of the place.",
                        "required": true,
                        "type": "string",
                        "format": "uuid"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "has permission.",
                        "schema": {
                            "properties": {
                                "hasPermission": {
                                    "type": "boolean",
                                    "example": false
                                }
                            },
                            "required": [
                                "hasPermission"
                            ]
                        }
                    }
                },
                "tags": [
                    "Place"
                ]
            }
        },
...
    },
...
    "definitions": {
        "Event": {
            "properties": {
                "@id": {
                    "type": "string",
                    "example": "http://culudb-silex.dev:8080/event/0823f57e-a6bd-450a-b4f5-8459b4b11043"
                },
                "name": {
                    "$ref": "#/definitions/TranslatableString"
                },
                "description": {
                    "$ref": "#/definitions/TranslatableString"
                },
                "available": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2015-05-07T12:02:53+00:00"
                },
                "image": {
                    "type": "string",
                    "example": "http://media.uitdatabank.be/20150416/153cfa0f-0d22-451e-bfd1-490b7c4ef109.jpg"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "example": "tagged"
                    }
                },
                "calendarSummary": {
                    "type": "string",
                    "example": "Every first day of the month"
                },
                "location": {
                    "$ref": "#/definitions/Place"
                },
                "organizer": {
                    "$ref": "#/definitions/Organizer"
                },
                "bookingInfo": {
                    "$ref": "#/definitions/BookingInfo"
                },
                "terms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Term"
                    }
                },
                "creator": {
                    "type": "string",
                    "example": "evenementen@stad.diksmuide.be"
                },
                "created": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2015-05-07T12:02:53+00:00"
                },
                "modified": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2015-05-07T12:02:53+00:00"
                },
                "publisher": {
                    "type": "string",
                    "example": "Invoerders Algemeen "
                },
                "endDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2015-05-07T12:02:53+00:00"
                },
                "startDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2015-05-07T12:02:53+00:00"
                },
                "calendarType": {
                    "type": "string",
                    "enum": [
                        "unknown",
                        "permanent",
                        "periodic",
                        "single"
                    ],
                    "example": "permanent"
                },
                "typicalAgeRange": {
                    "type": "string",
                    "example": "+18"
                },
                "performer": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Performer"
                    }
                },
                "sameAs": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "example": "http://www.uitinvlaanderen.be/agenda/e/zomerse-vrijdagen-den-engel/0823f57e-a6bd-450a-b4f5-8459b4b11043"
                    }
                },
                "seeAlso": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "example": "www.leuven.be"
                    }
                }
            },
            "required": [
                "@id",
                "name",
                "description",
                "creator",
                "created",
                "publisher",
                "calendarType"
            ]
        },
        "TranslatableString": {
            "properties": {
                "nl": {
                    "type": "string",
                    "example": "Nederlands"
                },
                "de": {
                    "type": "string",
                    "example": "Deutch"
                },
                "en": {
                    "type": "string",
                    "example": "English"
                },
                "fr": {
                    "type": "string",
                    "example": "Français"
                }
            },
            "required": [
                "nl"
            ]
        },
        "Address": {
            "properties": {
                "addressCountry": {
                    "type": "string",
                    "description": "The country code of the location.",
                    "example": "BE"
                },
                "addressLocality": {
                    "type": "string",
                    "description": "The locality name of the location.",
                    "example": "Leuven"
                },
                "postalCode": {
                    "type": "string",
                    "description": "The postalcode of the location.",
                    "example": 3000
                },
                "streetAddress": {
                    "type": "string",
                    "description": "The street name and number information of the location.",
                    "example": "Sluisstraat 79"
                }
            },
            "required": [
                "addressCountry",
                "addressLocality",
                "postalCode",
                "streetAddress"
            ]
        },
        "Term": {
            "properties": {
                "label": {
                    "type": "string",
                    "example": "Cycling"
                },
                "domain": {
                    "type": "string",
                    "example": "activities"
                },
                "id": {
                    "type": "string",
                    "example": "10.0.0.1"
                }
            },
            "required": [
                "label",
                "domain",
                "id"
            ]
        },
        "Performer": {
            "properties": {
                "name": {
                    "type": "string",
                    "description": "The name of the performer.",
                    "example": "Sindicato Sonico"
                }
            },
            "required": [
                "name"
            ]
        },
        "EventHistory": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/EventHistoryLog"
            }
        },
        "EventHistoryLog": {
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2015-05-07T12:02:53+00:00"
                },
                "description": {
                    "type": "string",
                    "example": "Aangemaakt in UDB2"
                },
                "author": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address of the author.",
                    "example": "johndoe@example.com"
                }
            },
            "required": [
                "date",
                "description"
            ]
        },
...
        "Place": {
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Or not to be."
                },
                "name": {
                    "type": "string",
                    "example": "This is the place to be"
                },
                "address": {
                    "$ref": "#/definitions/Address"
                },
                "bookingInfo": {
                    "$ref": "#/definitions/BookingInfo"
                },
                "terms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Term"
                    }
                }
            },
            "required": [
                "name",
                "address"
            ]
        },
        "Organizer": {
            "properties": {
                "name": {
                    "type": "string",
                    "example": "STUK"
                },
                "address": {
                    "$ref": "#/definitions/Address"
                },
                "email": {
                    "type": "string",
                    "format": "email",
                    "description": "An email address.",
                    "example": "info@stuk.be"
                },
                "phone": {
                    "type": "array",
                    "items": {
                        "type": "string",
                        "example": "016 320 300"
                    }
                }
            },
            "required": [
                "name"
            ]
        },
        "BookingInfo": {
            "properties": {
                "priceCurrency": {
                    "type": "string",
                    "example": "EUR"
                },
                "description": {
                    "type": "string",
                    "example": "No need to pay anything"
                },
                "name": {
                    "type": "string",
                    "example": "Free"
                },
                "price": {
                    "type": "number",
                    "format": "float",
                    "example": 0
                }
            },
            "required": [
                "price",
                "priceCurrency",
                "description"
            ]
        },
...
        "command": {
            "required": [
                "commandId"
            ],
            "properties": {
                "commandId": {
                    "type": "string",
                    "example": "8cdc13e62efaecb9d8c21d59a29b9de4"
                }
            }
        },
        "postObjectWithEventDescription": {
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is a description of a cultural event."
                }
            },
            "required": [
                "description"
            ]
        },
...
        "eventCreationData": {
            "properties": {
                "name": {
                    "type": "object",
                    "properties": {
                        "nl": {
                            "type": "string",
                            "example": "Super Awesome Party Night"
                        }
                    }
                },
                "type": {
                    "type": "object",
                    "description": "One of the event types available at: http://taxonomy.uitdatabank.be/api/domain/eventtype/classification",
                    "properties": {
                        "id": {
                            "type": "string",
                            "example": "0.50.4.0.0"
                        },
                        "label": {
                            "type": "string",
                            "example": "Concert"
                        },
                        "domain": {
                            "type": "string",
                            "enum": [
                                "eventtype"
                            ]
                        }
                    }
                },
                "location": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string",
                            "format": "uuid",
                            "description": "A universally unique identifier.",
                            "example": "6f072ba8-c510-40ac-b387-51f582650e27"
                        },
                        "name": {
                            "type": "string",
                            "description": "The name of the location.",
                            "example": "Versuz"
                        },
                        "address": {
                            "type": "object",
                            "properties": {
                                "addressCountry": {
                                    "type": "string",
                                    "example": "BE",
                                    "format": "ISO 3166-1 alpha-2"
                                },
                                "addressLocality": {
                                    "type": "string",
                                    "example": "Hasselt"
                                },
                                "postalCode": {
                                    "type": "string",
                                    "format": "postal code",
                                    "example": "3500"
                                },
                                "streetAddress": {
                                    "type": "string",
                                    "example": "Gouverneur Verwilghensingel 70",
                                    "format": "street name and number"
                                }
                            }
                        }
                    }
                },
                "calendarType": {
                    "type": "string",
                    "enum": [
                        "single",
                        "multiple",
                        "periodic",
                        "permanent"
                    ]
                },
                "openingHours": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "dayOfWeek": {
                                "type": "string",
                                "enum": [
                                    "maandag",
                                    "dinsdag",
                                    "woensdag",
                                    "donderdag",
                                    "vrijdag",
                                    "zaterdag",
                                    "zondag"
                                ]
                            },
                            "opens": {
                                "type": "string",
                                "example": "04:20"
                            },
                            "closes": {
                                "type": "string",
                                "example": "06:66"
                            }
                        }
                    }
                },
                "timestamps": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "date": {
                                "type": "string",
                                "format": "date",
                                "example": "2000-12-30"
                            },
                            "startHour": {
                                "type": "string",
                                "example": "04:20"
                            },
                            "endHour": {
                                "type": "string",
                                "example": "06:66"
                            }
                        }
                    }
                },
                "startDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2015-05-07T12:02:53+00:00"
                },
                "endDate": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2015-05-07T12:02:53+00:00"
                },
                "theme": {
                    "type": "object",
                    "description": "One of the available themes from http://taxonomy.uitdatabank.be/api/domain/theme/classification",
                    "properties": {
                        "id": {
                            "type": "string",
                            "example": "1.2.1.0.0"
                        },
                        "label": {
                            "type": "string",
                            "example": "Architectuur"
                        },
                        "domain": {
                            "type": "string",
                            "enum": [
                                "thema"
                            ],
                            "description": "This is hardcoded to 'thema' in the current application. It is passed to the API and then overwritten."
                        }
                    }
                }
            },
            "required": [
                "name",
                "type",
                "theme",
                "location",
                "calendarType",
                "startDate"
            ]
        },
...
        "addImageToOffer": {
            "properties": {
                "mediaObjectId": {
                    "type": "string",
                    "example": "73695986-e4cf-4b29-8699-13d7cd77af8c"
                }
            },
            "required": [
                "mediaObjectId"
            ]
        },
        "imageUpdateInfo": {
            "properties": {
                "description": {
                    "type": "string",
                    "description": "A description of the image."
                },
                "copyrightHolder": {
                    "type": "string",
                    "example": "Dirk Dirkington",
                    "description": "The name of the copyright holder."
                }
            },
            "required": [
                "description",
                "copyrightHolder"
            ]
        },
...
    }
}